requirements.txt
flask
python-dotenv
pydub
faster-whisper
google-genai
